Update Rotations/hunter_bm.lua
This is a small change to the BM rotation, as it 'should' cast cobra shot when focus is <=20, and it 'should' add the ability to cast explosive trap by pressing the shift key.

The thing I was not sure about was the number '6' in the local variable. I copied that from the Blood DK lua, so the number may be wrong. I have no idea how to figure out what the correct number is, but once that is determined (and tested), this behavior should be good to go.

Let me know if there is anything that you change, so that I can understand (and hopefully contribute more in the future. I am going to start working on a BM PvP set next.
JU1CYFRU1T